+++ Difi.Oppslagstjeneste.Klient.Tester/Envelope/PersonerEnvelopeTests.cs
@@ -1,9 +1,9 @@
 ﻿using Difi.Oppslagstjeneste.Klient.Domene.Entiteter.Enums;
 using Difi.Oppslagstjeneste.Klient.Envelope;
-using Difi.Oppslagstjeneste.Klient.Tests;
+using Difi.Oppslagstjeneste.Klient.Tests.Utilities;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Difi.Oppslagstjeneste.Klient.Tester.Envelope
+namespace Difi.Oppslagstjeneste.Klient.Tests.Envelope
 {
     [TestClass]
     public class PersonerEnvelopeTests
