+++ Difi.Oppslagstjeneste.Klient.Tester/Envelope/PrintSertifikatEnvelopeTests.cs
@@ -1,8 +1,8 @@
 ﻿using Difi.Oppslagstjeneste.Klient.Envelope;
-using Difi.Oppslagstjeneste.Klient.Tests;
+using Difi.Oppslagstjeneste.Klient.Tests.Utilities;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Difi.Oppslagstjeneste.Klient.Tester.Envelope
+namespace Difi.Oppslagstjeneste.Klient.Tests.Envelope
 {
     [TestClass]
     public class KonstruktørMethod : PersonerEnvelopeTests
