+++ Difi.Oppslagstjeneste.Klient.Tester/Integrasjon/Integrasjonstester.cs
@@ -5,7 +5,7 @@ using Difi.Oppslagstjeneste.Klient.Domene.Entiteter.Enums;
 using log4net.Config;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Difi.Oppslagstjeneste.Klient.Tester.Integrasjon
+namespace Difi.Oppslagstjeneste.Klient.Tests.Integrasjon
 {
     [TestClass]
     public class Integrasjonstester
@@ -16,7 +16,7 @@ namespace Difi.Oppslagstjeneste.Klient.Tester.Integrasjon
         public static void Init(TestContext context)
         {
             XmlConfigurator.Configure();
-            var resourceUtility = new ResourceUtility("Difi.Oppslagstjeneste.Klient.Tester.Ressurser.Sertifikater");
+            var resourceUtility = new ResourceUtility("Difi.Oppslagstjeneste.Klient.Tests.Ressurser.Sertifikater");
             var avsendersertifikat = new X509Certificate2(resourceUtility.ReadAllBytes(true, "DifiTestsertifikatKlient.p12"), "changeit", X509KeyStorageFlags.Exportable);
             var oppslagstjenesteKonfigurasjon = new OppslagstjenesteKonfigurasjon(Miljø.FunksjoneltTestmiljø, avsendersertifikat);
