+++ Difi.Oppslagstjeneste.Klient.Tester/Utilities/CompareObjects/IComparator.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Difi.Oppslagstjeneste.Klient.Tester.Utilities.CompareObjects
+namespace Difi.Oppslagstjeneste.Klient.Tests.Utilities.CompareObjects
 {
     public interface IComparator
     {
